Created CubeBackgrounModel class for containing useful functions to produce cube bg models.
